requirements.txt
aiohttp==3.8.3
asyncio==3.4.3
nest-asyncio==1.5.6
prometheus-async==22.2.0
prometheus-client==0.16.0
PyYAML==6.0
requests==2.28.2